[ADD] config default module import, move styles inside src
