+++ pages/_app.tsx
@@ -4,8 +4,8 @@ import { AppProps } from 'next/app';
 import { ThemeProvider } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
 import { CacheProvider, EmotionCache } from '@emotion/react';
-import theme from '../styles/theme';
-import createEmotionCache from '../src/utils/createEmotionCache';
+import createEmotionCache from '~/utils/createEmotionCache';
+import theme from '~/styles/theme';
 
 // Client-side cache, shared for the whole session of the user in the browser.
 const clientSideEmotionCache = createEmotionCache();
