+++ tsconfig.json
@@ -1,5 +1,12 @@
 {
   "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "~/components/*": ["src/components/*"],
+      "~/containers/*": ["src/containers/*"],
+      "~/utils/*": ["src/utils/*"],
+      "~/styles/*": ["src/styles/*"]
+    },
     "target": "es5",
     "lib": ["dom", "dom.iterable", "esnext"],
     "allowJs": true,
